Refactor resolving proc into legit class
